# cmd/app/main.go
package main

import (
	"context"
	"log/slog"
	"net/http"
	"os"
	"os/signal"
	"syscall"
	"time"

	"github.com/go-chi/chi/v5"

	favoriteservice "github.com/Oleja123/estate-agency/internal/application/favorite"
	imageservice "github.com/Oleja123/estate-agency/internal/application/image"
	propertyservice "github.com/Oleja123/estate-agency/internal/application/property"
	propertytypeservice "github.com/Oleja123/estate-agency/internal/application/property_type"
	"github.com/Oleja123/estate-agency/internal/application/token"
	userservice "github.com/Oleja123/estate-agency/internal/application/user"
	"github.com/Oleja123/estate-agency/internal/application/user/password"
	"github.com/Oleja123/estate-agency/internal/handler/auth"

	propertyhandler "github.com/Oleja123/estate-agency/internal/handler/property"
	propertytypehandler "github.com/Oleja123/estate-agency/internal/handler/property_type"
	userhandler "github.com/Oleja123/estate-agency/internal/handler/user"
	postgresqlclient "github.com/Oleja123/estate-agency/internal/infrastructure/client/postgresql"
	"github.com/Oleja123/estate-agency/internal/infrastructure/config"
	favoritedb "github.com/Oleja123/estate-agency/internal/infrastructure/favorite"
	"github.com/Oleja123/estate-agency/internal/infrastructure/filestore"
	geocoder "github.com/Oleja123/estate-agency/internal/infrastructure/geocoder"
	imagedb "github.com/Oleja123/estate-agency/internal/infrastructure/image"
	propertydb "github.com/Oleja123/estate-agency/internal/infrastructure/property"
	propertytypedb "github.com/Oleja123/estate-agency/internal/infrastructure/property_type"
	userdb "github.com/Oleja123/estate-agency/internal/infrastructure/user"

	swagger "github.com/Oleja123/estate-agency/internal/handler/swagger"
)

func main() {
	logger := slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelInfo}))

	cfgPath := os.Getenv("CONFIG_PATH")
	if cfgPath == "" {
		cfgPath = "config.yaml"
	}

	cfg, err := config.LoadConfig(cfgPath, logger)
	if err != nil {
		logger.Error("не удалось загрузить конфигурацию", "err", err, "path", cfgPath)
		os.Exit(1)
	}

	logger.Info("конфигурация загружена",
		"db_host", cfg.DbConfig.Host,
		"db_name", cfg.DbConfig.Database,
		"goose_path", cfg.GoosePath,
	)

	dbClient, err := postgresqlclient.NewClient(context.Background(), *logger, cfg)

	if err != nil {
		logger.Error("не удалось создать клиент базы данных", "err", err)
		os.Exit(1)
	}

	userStorage := userdb.New(dbClient, logger)
	propertyStorage := propertydb.New(dbClient, logger)
	favoriteStorage := favoritedb.New(dbClient, logger)
	imageStorage := imagedb.New(dbClient, logger)
	propertyTypeStorage := propertytypedb.New(dbClient, logger)

	tokSvc := token.NewMemoryService()
	userService := userservice.New(userStorage, logger, password.NewBcryptHasher(), tokSvc)
	propertyTypeService := propertytypeservice.New(propertyTypeStorage, logger)
	filestore := filestore.New("images/")
	favoriteService := favoriteservice.New(favoriteStorage, logger, filestore)

	var geoSvc geocoder.GeoService
	if cfg.GeoConfig.APIKey != "" {
		httpClient := &http.Client{Timeout: 10 * time.Second}
		geoSvc = geocoder.NewGeoapify(cfg.GeoConfig, httpClient)
	} else {
		geoSvc = geocoder.NewNoop()
	}

	if _, err := userService.SetUserRole(context.Background(), 1, "admin"); err != nil {
		logger.Error("не удалось установить начальную роль администратора", "err", err)
	}

	propertyService := propertyservice.New(propertyStorage, propertyTypeStorage, logger, geoSvc, favoriteService, filestore)
	imageService := imageservice.New(imageStorage, logger, filestore, "images/")

	router := chi.NewRouter()

	router.Use(func(next http.Handler) http.Handler {
		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
			w.Header().Set("Access-Control-Allow-Origin", "*")
			w.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, PATCH, DELETE, OPTIONS")
			w.Header().Set("Access-Control-Allow-Headers", "Content-Type, Authorization, X-Requested-With")
			w.Header().Set("Access-Control-Allow-Credentials", "true")
			if r.Method == http.MethodOptions {
				w.WriteHeader(http.StatusNoContent)
				return
			}
			next.ServeHTTP(w, r)
		})
	})

	auth.SetLogger(logger)

	authMw := auth.AuthMiddleware(tokSvc)

	uh := userhandler.NewUserHandler(userService, logger, favoriteService)
	uh.Register(router, "/users", authMw)

	auth.NewTokenHandler(tokSvc, userService).Register(router, "/tokens", authMw)

	ph := propertyhandler.NewPropertyHandler(propertyService, logger, favoriteService, imageService)
	ph.Register(router, "/properties", authMw)

	propertytypehandler.NewPropertyTypeHandler(propertyTypeService, logger).Register(router, "/property_types", authMw)

	swagger.RegisterSwagger(router, logger)

	port := os.Getenv("PORT")
	if port == "" {
		port = "8080"
	}

	srv := &http.Server{Addr: ":" + port, Handler: router}

	go func() {
		logger.Info("запуск сервера", "addr", srv.Addr)
		if err := srv.ListenAndServe(); err != nil && err != http.ErrServerClosed {
			logger.Error("сбой сервера", "err", err)
			os.Exit(1)
		}
	}()

	stop := make(chan os.Signal, 1)
	signal.Notify(stop, syscall.SIGINT, syscall.SIGTERM)
	<-stop

	ctxShut, cancel := context.WithTimeout(context.Background(), 5*time.Second)
	defer cancel()
	logger.Info("завершение работы сервера")
	if err := srv.Shutdown(ctxShut); err != nil {
		logger.Error("ошибка при завершении работы сервера", "err", err)
		os.Exit(1)
	}

}