# docs/swagger.yaml
swagger: "2.0"
info:
  title: Estate Agency API
  description: 'Введите токен в формате: "Bearer <token>" (включая слово Bearer и пробел)'
  version: 0.0.1
host: localhost:8080
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
security:
  - BearerAuth: []
paths:
...
  /properties:
    get:
      summary: List properties
      tags: [properties]
      description: Returns paginated list of properties. Requires authentication.
      parameters:
        - name: limit
          in: query
          type: integer
          example: 10
        - name: offset
          in: query
          type: integer
          example: 0
        - name: ids
          in: query
          type: string
          example: "1,2,3"
        - name: type_ids
          in: query
          type: string
          example: "2,3"
        - name: type_id
          in: query
          type: integer
          example: 2
        - name: transaction_type
          in: query
          type: string
          example: "sale"
        - name: city
          in: query
          type: string
          example: "New York"
        - name: property_status
          in: query
          type: string
          example: "available"
        - name: created_by
          in: query
          type: integer
          example: 1
        - name: min_price
          in: query
          type: number
          example: 100000
        - name: max_price
          in: query
          type: number
          example: 500000
        - name: min_area
          in: query
          type: number
          example: 50
        - name: max_area
          in: query
          type: number
          example: 200
        - name: search
          in: query
          type: string
          example: "spacious"
        - name: latitude
          in: query
          type: number
          example: 40.7128
        - name: longitude
          in: query
          type: number
          example: -74.006
        - name: radius_km
          in: query
          type: number
          example: 5
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/propertyhandler.ListPropertiesResponseDoc'
          examples:
            application/json:
              properties:
                - id: 1
                  title: "Luxury Apartment"
                  property_description: "A spacious 3-bedroom apartment."
                  type_id: 2
                  transaction_type: "sale"
                  price: 350000
                  area: 120
                  property_address: "123 Main Street"
                  latitude: 40.7128
                  longitude: -74.006
                  city: "New York"
                  property_status: "available"
                  images:
                    - filename: "photo1.jpg"
                      data: "base64encodedstring=="
                  image:
                    filename: "photo1.jpg"
                    data: "base64encodedstring=="
                  created_at: "2025-11-29T12:00:00Z"
                  updated_at: "2025-11-29T12:00:00Z"
              total: 1
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/ErrorResponse'
    post:
      summary: Create property
      tags: [properties]
      description: |
        Create a new property. Requires admin role.

        Адрес (`property_address`) должен быть корректным и непустым — при создании сервиса будут получены координаты (``latitude``, ``longitude``) с помощью Geoapify Geocoding API.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/propertyhandler.PropertyCreateDoc'
      responses:
        "201":
          description: Создано
          schema:
            $ref: '#/definitions/propertyhandler.PropertyDTODoc'
        "400":
          description: Некорректный запрос
        "401":
          description: Требуется авторизация
        "403":
          description: Доступ запрещён
        "409":
          description: Конфликт
        "422":
          description: Unprocessable Entity — геокодирование не вернуло координаты
          schema:
            $ref: '#/definitions/GeocodingErrorResponse'
        "502":
          description: Bad Gateway — ошибка внешнего геосервиса
          schema:
            $ref: '#/definitions/GeocodingErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/ErrorResponse'
...
  /property_types:
    get:
      summary: List property types
      tags: [property_types]
      parameters:
        - name: limit
          in: query
          type: integer
          example: 10
        - name: offset
          in: query
          type: integer
          example: 0
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/propertytypehandler.ListPropertyTypesResponseDoc'
        "401":
          description: Требуется авторизация
        "500":
          description: Внутренняя ошибка сервера
    post:
      summary: Create property type
      tags: [property_types]
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/propertytypehandler.CreatePropertyTypeDoc'
      responses:
        "201":
          description: Создано
          schema:
            $ref: '#/definitions/propertytypehandler.PropertyTypeDTODoc'
        "400":
          description: Некорректный запрос
        "401":
          description: Требуется авторизация
        "403":
          description: Доступ запрещён
        "409":
          description: Конфликт
        "500":
          description: Внутренняя ошибка сервера
...
  /users:
    get:
      summary: List users
      tags: [users]
      parameters:
        - name: limit
          in: query
          type: integer
          example: 10
        - name: offset
          in: query
          type: integer
          example: 0
        - name: ids
          in: query
          type: string
          example: "1,2,3"
        - name: email
          in: query
          type: string
          example: "user@example.com"
        - name: role
          in: query
          type: string
          example: "client"
        - name: is_active
          in: query
          type: boolean
          example: true
        - name: search
          in: query
          type: string
          example: "john"
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userhandler.ListUsersResponseDoc'
        "401":
          description: Требуется авторизация
        "500":
          description: Внутренняя ошибка сервера
...
  /tokens/refresh:
    post:
      summary: Refresh token
      tags: [tokens]
      security: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/auth.RefreshRequestDoc'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userhandler.LoginResponseDoc'
          examples:
            application/json:
              user:
                id: 1
                email: "user@example.com"
                first_name: "John"
                last_name: "Doe"
                role: "user"
                is_active: true
                created_at: "2025-11-29T12:00:00Z"
                updated_at: "2025-11-29T12:00:00Z"
              access_token: "eyJhbGciOiJIUzI1..."
              refresh_token: "dGhpc2lzYXJlZnJlc2h0b2tlbg=="
              expires_at: "2025-11-29T14:00:00Z"
        "400":
          description: Некорректный запрос
        "401":
          description: Недействительный токен обновления
  /tokens/logout:
    post:
      summary: Logout (invalidate refresh token)
      tags: [tokens]
      security: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/auth.RefreshRequestDoc'
      responses:
        "200":
          description: OK
          schema:
            type: object
            properties:
              message: { type: string, example: "выход выполнен" }
        "400":
          description: Некорректный запрос
        "401":
          description: Недействительный токен обновления
        "500":
          description: Внутренняя ошибка сервера

# -------------------- Definitions --------------------
definitions:
  propertyhandler.PropertyDTODoc:
    type: object
    properties:
      id: { type: integer, example: 1 }
      title: { type: string, example: "Luxury Apartment" }
      property_description: { type: string, example: "A spacious 3-bedroom apartment." }
      type_id: { type: integer, example: 2 }
      transaction_type: { type: string, example: "sale" }
      price: { type: number, example: 350000 }
      area: { type: number, example: 120 }
      property_address: { type: string, example: "123 Main Street", description: "Должен быть корректным, непустым адрес. Координаты (latitude, longitude) вычисляются с помощью Geoapify Geocoding API при создании/обновлении." }
      latitude: { type: number, example: 40.7128 }
      longitude: { type: number, example: -74.006 }
      city: { type: string, example: "New York" }
      property_status: { type: string, example: "available" }
      is_favorited: { type: boolean, example: false }
      created_at: { type: string, format: date-time, example: "2025-11-29T12:00:00Z" }
      updated_at: { type: string, format: date-time, example: "2025-11-29T12:00:00Z" }
      images:
        type: array
        items:
          $ref: '#/definitions/imagedto.ImageFileDoc'
      image:
        $ref: '#/definitions/imagedto.ImageFileDoc'
  propertyhandler.PropertyCreateDoc:
    type: object
    required:
      - title
      - property_description
      - price
      - area
      - city
      - property_address
      - transaction_type
      - type_id
    properties:
      title: { type: string, example: "Luxury Apartment" }
      property_description: { type: string, example: "A spacious 3-bedroom apartment." }
      type_id: { type: integer, example: 2 }
      transaction_type: { type: string, example: "sale" }
      price: { type: number, example: 350000 }
      area: { type: number, example: 120 }
      property_address: { type: string, example: "123 Main Street", description: "Должен быть корректным, непустым адрес. Координаты (latitude, longitude) вычисляются с помощью Geoapify Geocoding API при создании/обновлении." }
      city: { type: string, example: "New York" }
...
  propertyhandler.ListPropertiesResponseDoc:
    type: object
    properties:
      properties:
        type: array
        items:
          $ref: '#/definitions/propertyhandler.PropertyDTODoc'
      total: { type: integer, example: 42 }

  # Image DTOs removed
  imagedto.ImageFileDoc:
    type: object
    properties:
      filename: { type: string, example: "photo1.jpg" }
      data: { type: string, example: "base64encodedstring==", description: "Base64 encoded binary data of the image" }
...
  propertytypehandler.PropertyTypeDTODoc:
    type: object
    properties:
      id: { type: integer, example: 1 }
      name: { type: string, example: "Apartment" }
      created_at: { type: string, format: date-time, example: "2025-11-29T12:00:00Z" }
  propertytypehandler.ListPropertyTypesResponseDoc:
    type: object
    properties:
      types:
        type: array
        items:
          $ref: '#/definitions/propertytypehandler.PropertyTypeDTODoc'
      total: { type: integer, example: 3 }
  propertytypehandler.CreatePropertyTypeDoc:
    type: object
    required:
      - name
    properties:
      name: { type: string, example: "Apartment" }
...
  userhandler.PublicUserDoc:
    type: object
    properties:
      id: { type: integer, example: 1 }
      email: { type: string, example: "user@example.com" }
      first_name: { type: string, example: "John" }
      last_name: { type: string, example: "Doe" }
      phone_number: { type: string, example: "+1234567890", nullable: true }
      role: { type: string, example: "user" }
      is_active: { type: boolean, example: true }
      created_at: { type: string, format: date-time, example: "2025-11-29T12:00:00Z" }
      updated_at: { type: string, format: date-time, example: "2025-11-29T12:00:00Z" }
...
  userhandler.LoginResponseDoc:
    type: object
    properties:
      user:
        $ref: '#/definitions/userhandler.PublicUserDoc'
      access_token: { type: string, example: "eyJhbGciOiJIUzI1..." }
      refresh_token: { type: string, example: "dGhpc2lzYXJlZnJlc2h0b2tlbg==" }
      expires_at: { type: string, format: date-time, example: "2025-11-29T14:00:00Z" }
...
  userhandler.ListUsersResponseDoc:
    type: object
    properties:
      users:
        type: array
        items:
          $ref: '#/definitions/userhandler.PublicUserDoc'
      total: { type: integer, example: 5 }
  auth.RefreshRequestDoc:
    type: object
    properties:
      refresh_token: { type: string, example: "dGhpc2lzYXJlZnJlc2h0b2tlbg==" }
  ErrorResponse:
    type: object
    properties:
      code: { type: integer, example: 400 }
      message: { type: string, example: "Некорректные параметры запроса" }
      details:
        type: object
        additionalProperties:
          type: string
  GeocodingErrorResponse:
    type: object
    properties:
      error: { type: string, example: "ошибка геокодирования: external service error" }
      provider: { type: string, example: "geoapify" }
      code: { type: string, example: "request" }
      details: { type: string, example: "timeout while connecting to geo service" }
      address: { type: string, example: "123 Main Street" }
      status_code: { type: integer, example: 504, nullable: true }
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header